% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_tools.R
\name{Cir_area}
\alias{Cir_area}
\title{Circle area calculations}
\usage{
Cir_area(rad)
}
\arguments{
\item{rad}{The radius of the circle}
}
\value{
The Area of the circle
}
\description{
Calculate the area of a circle with the given radius
}
\examples{
temp1 <- Cir_area(20);
temp2 <- Cir_area( c(12, 10, 33) );
}
